requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
authlib==1.2.1
python-dotenv==1.0.0
pydantic==1.10.16
python-multipart==0.0.6